Added Authintcation & product ( create , Update , Delete)

+++ database/migrations/2021_10_27_111523_create_reviews_table.php
@@ -15,7 +15,7 @@ class CreateReviewsTable extends Migration
     {
         Schema::create('reviews', function (Blueprint $table) {
             $table->id();
-            $table->integer('product_id')->unsigned()->index();
+            $table->bigInteger('product_id')->unsigned()->index();
             $table->foreign('product_id')->references('id')->on('products')->onDelete('cascade');
             $table->string('customer');
             $table->text('review');
@@ -24,7 +24,7 @@ class CreateReviewsTable extends Migration
         });
     }
 
- 
+
     public function down()
     {
         Schema::dropIfExists('reviews');
